SharingVisitor final stricture.
