+++ src/Search/Querying/SharingNormalizerVisitor.cs
@@ -6,13 +6,11 @@ using SenseNet.Search.Querying.Parser.Predicates;
 
 namespace SenseNet.Search.Querying
 {
-    //UNDONE:<? Move SharingVisitor to ContentRepository (?)
-
     /// <summary>
     /// Normalizes the predicate tree. Removes irrelevant terma and unnecessary parentheses.
     /// Only works correctly if the predicate tree does not contain negative logical clause.
     /// </summary>
-    internal class SharingNormalizerVisitor : SnQueryVisitor
+    internal class SharingNormalizerVisitor : SnQueryVisitor //UNDONE:<? Move to ContentRepository (?)
     {
         public override LogicalClause VisitLogicalClause(LogicalClause clause)
         {
